Dev (#47)
* Prevent adding the same component twice (#46)

* Prevent adding the same component twice

* Fix wording

* Minor fixes

* Better Framebuffers (#48)

* EditorCamera (#49)

* Editor 3 (#52)

* 2D Physics (#53)

* WIP

* Add box2d submodule

* WIP

* Modify box2d

* Remove original box2d

* Add forked version of box2d as submodule

* WIP

* GUID (#56)

Add GUID

* Update readme

* Add submodules

* Add test project

* Move engine code to separate folder within repository

* Move engine third parties to separate folder

* Move editor third_party to separate folder

* WIP

* Removed cpp-comon-lib submodule

* Change shadec lib to shared

* Move gh pages to separate branch and add deploy action

+++ samples/hello_cube/main.cpp
@@ -100,7 +100,7 @@ public:
 private:
     bool onWindowResize(jng::WindowResizeEvent& event)
     {
-        m_camera.setViewportSize(event.getWidth(), event.getHeight());
+        m_camera.setViewportSize(static_cast<float>(event.getWidth()), static_cast<float>(event.getHeight()));
         m_cameraUBO->setData(glm::value_ptr(m_camera.getVP(glm::mat4{ 1.f })), sizeof(glm::mat4), 0);
         return false;
     }
